+++ packages/sdk/src/order/fill-order/crypto-punks.ts
@@ -14,7 +14,6 @@ import { getRequiredWallet } from "../../common/get-required-wallet"
 import { invertOrder } from "./invert-order"
 import type { CryptoPunksOrderFillRequest, OrderFillSendData, OrderHandler } from "./types"
 import { getUpdatedCalldata } from "./common/get-updated-call"
-import { hexifyOptionsValue } from "./common/hexify-options-value"
 
 export class CryptoPunksOrderHandler implements OrderHandler<CryptoPunksOrderFillRequest> {
 	constructor(
@@ -44,20 +43,16 @@ export class CryptoPunksOrderHandler implements OrderHandler<CryptoPunksOrderFil
 	async getTransactionData(
 		initial: SimpleCryptoPunkOrder, inverted: SimpleCryptoPunkOrder,
 	): Promise<OrderFillSendData> {
-		const options = hexifyOptionsValue({
+		const options = {
 			...this.getMatchV2Options(initial, inverted),
 			additionalData: getUpdatedCalldata(this.sdkConfig),
-		})
+		}
 		const functionCall = this.getPunkOrderCallMethod(initial)
 		await functionCall.estimateGas({
 			from: await getRequiredWallet(this.ethereum).getFrom(),
 			value: options.value,
 		})
-
-		return {
-			functionCall,
-			options,
-		}
+		return { functionCall, options }
 	}
 
 	async sendTransaction(