+++ CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.86](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.85...v0.12.86) (2022-08-17)
+
+
+### Bug Fixes
+
+* tx value converter ([#182](https://github.com/rarible/protocol-ethereum-sdk/issues/182)) ([28c1413](https://github.com/rarible/protocol-ethereum-sdk/commit/28c1413c286b5009485ca80f93d7fdb08c5557ac))
+
+
+
+
+
+## [0.12.85](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.84...v0.12.85) (2022-08-17)
+
+
+### Bug Fixes
+
+* batch purchase calldata ([ff1d474](https://github.com/rarible/protocol-ethereum-sdk/commit/ff1d4742db5c10d55306c368ff8780daf31d0baf))
+
+
+
+
+
+## [0.12.84](https://github.com/rarible/ethereum-sdk/compare/v0.12.83...v0.12.84) (2022-08-17)
+
+
+### Features
+
+* batch purchase support update ([#179](https://github.com/rarible/ethereum-sdk/issues/179)) ([c91898d](https://github.com/rarible/ethereum-sdk/commit/c91898d59da5a1739bec4d9a9de8ff8879c1da2c))
+
+
+
+
+
 ## [0.12.83](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.82...v0.12.83) (2022-08-16)